parte dello score aggiornata

+++ Backend/src/main/java/com/sharingnotes/MongoDb/MongoDb.java
@@ -47,7 +47,7 @@ public class MongoDb {
     public void updateScore(Document document,String collection,int score){
         int newNVoti=(document.getInteger("nvoti")+1);
         int newSomma=(document.getInteger("sommaVoti")+score);
-        int media = Math.round(newSomma/ newNVoti);
+        float media = Math.round(newSomma/ newNVoti);
         database.getCollection(collection).updateOne(Filters.eq("_id",document.get("_id")), Updates.set("nvoti",newNVoti));
         database.getCollection(collection).updateOne(Filters.eq("_id",document.get("_id")), Updates.set("sommaVoti",newSomma));
         database.getCollection(collection).updateOne(Filters.eq("_id",document.get("_id")), Updates.set("rating",media));
